--- frontend/src/libs/packages/socket/socket.service.ts
@@ -1,26 +1,0 @@
-import { type Socket } from 'socket.io-client';
-
-import { ClientSocketEvent } from '~/libs/packages/socket/libs/enums/enum.js';
-
-import { socket } from './socket.js';
-
-class SocketService {
-  private io: Socket;
-
-  public constructor() {
-    this.io = socket.getInstance();
-  }
-
-  public addListener(
-    event: keyof typeof ClientSocketEvent,
-    listener: () => void,
-  ): void {
-    this.io.on(ClientSocketEvent[event], listener);
-  }
-
-  public disconnect(): void {
-    this.io.disconnect();
-  }
-}
-
-export { SocketService };
